added common cmakelists.txt

+++ Difficult_numbes.cpp
@@ -23,11 +23,11 @@ int sum(int i) {
 int f() {
     vector<size_t> vec(1'000000, 0);
 
-    for(size_t i = 1; i < 1'000'000; ++i) {
+    for(int i = 1; i < 1'000'000; ++i) {
         double k = (3.0 * i / (sum(i) * sum(i)));
 
-        if(k - int(k) == 0)
-            vec[k] = i;
+        if(k - static_cast<int>(k) == 0)
+            vec[static_cast<int>(k)] = i;
 
     }
